Move the OSInfo code to its own file.
